+++ src/constants.ts
@@ -1,4 +1,5 @@
 /**
+ * 
  * Copyright 2024-present Coinbase Global, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");